refactor(modules): Change to a new structure
